+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 > only (ignore internal cleanup) one-line per change Ignore spec documents, and
 > deprioritize test-only changes
 
-## v0.8.7 2026-08-12
+## v0.8.7 2026-08-13
 
 - Return the assigned value when reassigning variables and preserve it in the
   surrounding expression.